pyproject.toml
[project]
name = "granachat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
